pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "MindScope AI - Advanced AI Assistant powered by Gemini Pro"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.42.2",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "uuid>=1.30",
    "pyttsx3>=2.90",
    "SpeechRecognition>=3.10.0",
    "gtts>=2.5.4",
    "numpy>=1.24.0",
    "pandas>=2.2.3",
]

[project.URLs]
Homepage = "https://github.com/Riaz-Hussain-Saifi/Python-Chat-Bot"